+++ src/main/java/com/arcanc/pulselib/content/event/ClientEvents.java
@@ -66,23 +66,24 @@ public class ClientEvents
 	/*private static void registerClientExtensions(final RegisterSpecialModelRendererEvent event)
 	{
 		event.register(PLibDatabase.rl("test_block"), TestBlockItemRenderer.Unbaked.MAP_CODEC);
-	}
+	}*/
 	
-	private static void registerRenderers(final EntityRenderersEvent.RegisterRenderers event)
+	/*private static void registerRenderers(final EntityRenderersEvent.RegisterRenderers event)
 	{
 		event.registerBlockEntityRenderer(Registration.BETypeReg.TEST_BLOCK_ENTITY.get(), TestBlockEntityRenderer :: new);
 		event.registerEntityRenderer(Registration.EntityTypeReg.TEST_ENTITY.get(), TestEntityRender :: new);
-	}
+	}*/
 	
-	private static void registerCustomTextures(final CustomEvents.PLibRegisterTextureEvent event)
+	/*private static void registerCustomTextures(final CustomEvents.PLibRegisterTextureEvent event)
 	{
+		event.addTextureLocation(TestEntityRender.SPHERE).
+				addTextureLocation(TestEntityRender.TUBE).
+				addTextureLocation(TestEntityRender.TORUS).
+				addTextureLocation(TestEntityRender.ZERO);
 		event.addTextureLocation(TestBlockEntityRenderer.CUBE).
 				addTextureLocation(TestBlockEntityRenderer.TORUS).
 				addTextureLocation(TestBlockEntityRenderer.TUBE).
 				addTextureLocation(TestBlockEntityRenderer.PYRAMID);
-		event.addTextureLocation(TestEntityRender.SPHERE).
-				addTextureLocation(TestEntityRender.TUBE).
-				addTextureLocation(TestEntityRender.TORUS);
 		event.addTextureLocation(TestBlockItemRenderer.PYRAMID).
 				addTextureLocation(TestBlockItemRenderer.CIRCLE);
 	}*/