+++ src/main/java/com/arcanc/pulselib/content/renderer/PEntityRenderLayer.java
@@ -1,0 +1,127 @@
+/**
+ * @author ArcAnc
+ * Created at: 20.05.2026
+ * Copyright (c) 2026
+ * <p>
+ * This code is licensed under "Arc's License of Common Sense"
+ * Details can be found in the license file in the root folder of this project
+ */
+
+package com.arcanc.pulselib.content.renderer;
+
+
+import com.arcanc.pulselib.content.animatable.PAnimatable;
+import com.arcanc.pulselib.content.animatable.PAnimationController;
+import com.arcanc.pulselib.content.model.baked.PBakedBone;
+import com.arcanc.pulselib.content.model.baked.PBakedMesh;
+import com.arcanc.pulselib.content.model.baked.PBakedModel;
+import com.arcanc.pulselib.content.renderer.base.PEntityRenderState;
+import com.arcanc.pulselib.content.renderer.modelData.PModelData;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.renderer.SubmitNodeCollector;
+import net.minecraft.client.renderer.entity.state.EntityRenderState;
+import net.minecraft.client.renderer.rendertype.RenderType;
+import net.minecraft.client.renderer.state.level.CameraRenderState;
+import net.minecraft.resources.Identifier;
+import net.minecraft.world.entity.Entity;
+import org.joml.Quaternionf;
+import org.joml.Vector3f;
+import org.jspecify.annotations.Nullable;
+
+import java.util.Collection;
+import java.util.function.Function;
+
+public abstract class PEntityRenderLayer<T extends Entity & PAnimatable<T>, RS extends EntityRenderState & PEntityRenderState<T>>
+{
+	private final PModelData modelData;
+	private final Function<Identifier, RenderType> renderType;
+	
+	public PEntityRenderLayer(PModelData modelData, Function<Identifier, RenderType> renderType)
+	{
+		this.modelData = modelData;
+		this.renderType = renderType;
+	}
+	
+	public PModelData getModelData(RS renderState)
+	{
+		return this.modelData;
+	}
+	
+	public @Nullable PBakedModel getModel(RS renderState)
+	{
+		return getModelData(renderState).getModel();
+	}
+	
+	public RenderType getRenderType(Identifier texture)
+	{
+		return this.renderType.apply(texture);
+	}
+	
+	public boolean shouldRender(RS renderState)
+	{
+		return true;
+	}
+	
+	public int getColor(RS renderState, PBakedBone bone, PBakedMesh mesh, int packedColor)
+	{
+		return packedColor;
+	}
+	
+	public int getPackedLight(RS renderState, int packedLight)
+	{
+		return packedLight;
+	}
+	
+	public int getPackedOverlay(RS renderState, int packedOverlay)
+	{
+		return packedOverlay;
+	}
+	
+	public void submit(PEntityRenderer<T, RS> renderer,
+	                   RS renderState,
+	                   PoseStack poseStack,
+	                   SubmitNodeCollector submitNodeCollector,
+	                   CameraRenderState cameraRenderState,
+	                   Collection<PAnimationController<T>> controllers,
+	                   int packedColor,
+	                   int packedLight,
+	                   int packedOverlay,
+	                   float partialTick)
+	{
+		PBakedModel model = getModel(renderState);
+		if (model == null || !shouldRender(renderState))
+			return;
+		
+		PModelData data = getModelData(renderState);
+		for (PBakedBone bone : model.bones())
+			renderer.perBoneSubmit(
+					renderState,
+					poseStack,
+					bone,
+					controllers,
+					data,
+					this :: getRenderType,
+					packedColor,
+					getPackedLight(renderState, packedLight),
+					getPackedOverlay(renderState, packedOverlay),
+					partialTick,
+					submitNodeCollector,
+					cameraRenderState,
+					this);
+	}
+	
+	public Vector3f offset()
+	{
+		return new Vector3f();
+	}
+	
+	public Quaternionf rotation()
+	{
+		return new Quaternionf();
+	}
+	
+	public Vector3f scale()
+	{
+		return new Vector3f(1, 1, 1);
+	}
+}